Déplace des types
